+++ views/assets/JS/socios_comerciales.js
@@ -93,7 +93,7 @@ $(document).ready(function() {
                 render: function (data, type, row) { // 'row' contiene toda la fila de datos
                     return `<div class="text-center">
                                 <button class="btn btn-xs editar-btn" data-row='${JSON.stringify(row)}'><i class="fa fa-pencil"></i></button>
-                                <button class="btn btn-xs eliminar-btn" data-pk="${row[15]}"><i class="fa fa-trash"></i></button>
+                                <button class="btn btn-xs eliminar-btn" data-pk="${row[15]}" data-nombre="${row[0]}"><i class="fa fa-trash"></i></button>
                             </div>`;
                 }
             }
@@ -126,13 +126,21 @@ $(document).ready(function() {
     $('#socioTable').on('click', '.eliminar-btn', function () {
 
         const pkSocioComercial = $(this).data('pk');
+        const nombreSocio = $(this).data('nombre');
 
-        var modal = $('[data-remodal-id="remodal"]').remodal();
-
-        modal.open();
-
-        $(document).on("confirmation", ".remodal", function () {
-            eliminarSocioComercial(pkSocioComercial);    
+        Swal.fire({
+            title: `¿Eliminar a ${nombreSocio}?`,
+            text: "No se podrá recuperar",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#B71C1C",
+            cancelButtonColor: "#C1C0C0",
+            confirmButtonText: "Eliminar",
+            cancelButtonText: "Cancelar"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                eliminarSocioComercial(pkSocioComercial);    
+            }
         });
         
     });
