+++ views/assets/JS/usuarios.js
@@ -58,7 +58,7 @@ $(document).ready(function() {
                 render: function (data, type, row) { // 'row' contiene toda la fila de datos
                     return `<div class="text-center">
                                 <button class="btn btn-xs editar-btn" data-row='${JSON.stringify(row)}'><i class="fa fa-pencil"></i></button>
-                                <button class="btn btn-xs eliminar-btn" data-pk="${row[4]}"><i class="fa fa-trash"></i></button>
+                                <button class="btn btn-xs eliminar-btn" data-pk="${row[4]}" data-nombre="${row[0]}"><i class="fa fa-trash"></i></button>
                             </div>`;
                 }
             }
@@ -87,16 +87,24 @@ $(document).ready(function() {
     });
 
     // Eliminar
-    $('#usuarioTable').one('click', '.eliminar-btn', function () {
+    $('#usuarioTable').on('click', '.eliminar-btn', function () {
 
         const pkUsuario = $(this).data('pk');
+        const nombreUsuario = $(this).data('nombre');
 
-        var modal = $('[data-remodal-id="remodal"]').remodal();
-
-        modal.open();
-
-        $(document).on("confirmation", ".remodal", function () {
-            eliminarUsuario(pkUsuario);    
+        Swal.fire({
+            title: `¿Eliminar a ${nombreUsuario}?`,
+            text: "No se podrá recuperar",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#B71C1C",
+            cancelButtonColor: "#C1C0C0",
+            confirmButtonText: "Eliminar",
+            cancelButtonText: "Cancelar"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                eliminarUsuario(pkUsuario);  
+            }
         });
         
     });
